handle error packet in backend handshake phase to avoid ArrayIndexOutOfBoundException
